tests/wpt/tests/css/css-transforms/animation/transform-interpolation-007.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>transform interpolation</title>
<link rel="help" href="https://drafts.csswg.org/css-transforms-1/#matrix-interpolation">
<meta name="assert" content="transform does not interpolate between non-invertible and invertible matrices">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
  <template id="target-template">
    <div></div>
 </template>
</body>

<script>
test_no_interpolation({
  property: 'transform',
  from: 'matrix3d(2, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1)',
  to: 'matrix(3, 0, 0, 3, 0, 0)',
});
</script>
